v2/solver/upgrade: Fix replaces of provided pkg
When is missing the conflicts in the new package
that provides an old package that is saw as upgradable
by the provided and at the same time the package
is yet installed and upgrade.

Example: the package app-misc/c_rehash-1.7 is
candidated for the upgrade with the
dev-libs/openssl-1.1.1q that provides
app-misc/c_rehash-1.8.

In the previous implementation the candidates
doesn't add app-misc/c_rehash as package to
remove.

This patch add a check for any provides
candidated for the upgrade and add the
existing packages replaced to the remove
queue.