Rewrite the `luet-build create-repo` command
This new implementation now use `treev2` key
on repository.yaml with a tarball that contains
already the metadata.json files and the provides.yaml.

The docker repository is no more supported.
This feature will be restored when will be rewritten
the `luet-build build` command.

Bump repositories with this new implementation
will break old clients. Will be needed to upgrade
luet before upgrade the system.